+++ src/containerapp/azext_containerapp/_utils.py
@@ -1389,10 +1389,7 @@ def patch_new_custom_domain(cmd, resource_group_name, name, new_custom_domains):
         r = ContainerAppClient.update(cmd, resource_group_name, name, envelope)
     except CLIError as e:
         handle_raw_exception(e)
-    if "customDomains" in r["properties"]["configuration"]["ingress"]:
-        return list(r["properties"]["configuration"]["ingress"]["customDomains"])
-    else:
-        return []
+    return safe_get(r, "properties", "configuration", "ingress", "customDomains", default=[])
 
 
 def get_custom_domains(cmd, resource_group_name, name, location=None, environment=None):
@@ -1404,10 +1401,7 @@ def get_custom_domains(cmd, resource_group_name, name, location=None, environmen
                 raise ResourceNotFoundError('Container app {} is not in location {}.'.format(name, location))
         if environment and (_get_name(environment) != _get_name(app["properties"]["managedEnvironmentId"])):
             raise ResourceNotFoundError('Container app {} is not under environment {}.'.format(name, environment))
-        if "ingress" in app["properties"]["configuration"] and "customDomains" in app["properties"]["configuration"]["ingress"]:
-            custom_domains = app["properties"]["configuration"]["ingress"]["customDomains"]
-        else:
-            custom_domains = []
+        custom_domains = safe_get(app, "properties", "configuration", "ingress", "customDomains", default=[])
     except CLIError as e:
         handle_raw_exception(e)
     return custom_domains