+++ src/containerapp/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.3.20
+++++++
+* Fix custom domain null issue for `az containerapp hostname list` and `az containerapp hostname delete` command
+
 0.3.19
 ++++++
 * Fix "'NoneType' object is not iterable" error in `az containerapp hostname bind` command